chore(xlsx): extract chart title + manual layout primitives into chart/title and chart/layout
Move every parseTitle / buildTitle / buildTitleSpPr / resolveTitle / normalizeTitle
helper from chart-reader.ts and chart-writer.ts into chart/title.ts. Lift
ResolvedManualLayout / normalizeManualLayout / normalizeLayoutCoordinate /
buildManualLayout into chart/layout.ts so every per-host module can reuse the
shared writer-side layout primitives without duplication.

Drop the chart-clone.ts copies of the title normalizers and the local
normalizeLayoutCoordinate; chart-clone now imports the shared helpers
directly. Clone-side resolveTitle* overrides (3-arg source/override
signatures) stay in chart-clone.ts because they are shape-incompatible
with the writer-side single-arg resolvers but they delegate to the
shared chart/title normalizers.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
