+++ src/xlsx/chart-writer.ts
@@ -46,6 +46,45 @@ import {
   buildSideWallThickness,
   buildView3D,
 } from "./chart/walls";
+import {
+  type ResolvedManualLayout,
+  buildManualLayout,
+  normalizeManualLayout,
+} from "./chart/layout";
+import {
+  FONT_SIZE_MAX_PT,
+  FONT_SIZE_MIN_PT,
+  FONT_SZ_PER_POINT,
+  ROTATION_MAX_DEG,
+  ROTATION_MIN_DEG,
+  TXPR_ROT_PER_DEGREE,
+} from "./chart/text";
+import {
+  buildTitle,
+  buildTitleSpPr,
+  normalizeTitleBold,
+  normalizeTitleColor,
+  normalizeTitleFontFamily,
+  normalizeTitleFontSize,
+  normalizeTitleItalic,
+  normalizeTitleRotation,
+  normalizeTitleStrike,
+  normalizeTitleUnderline,
+  resolveTitleBold,
+  resolveTitleBorderColor,
+  resolveTitleBorderDash,
+  resolveTitleBorderWidth,
+  resolveTitleColor,
+  resolveTitleFillColor,
+  resolveTitleFontFamily,
+  resolveTitleFontSize,
+  resolveTitleItalic,
+  resolveTitleLayout,
+  resolveTitleOverlay,
+  resolveTitleRotation,
+  resolveTitleStrike,
+  resolveTitleUnderline,
+} from "./chart/title";
 
 // ── Namespaces ───────────────────────────────────────────────────────
 
@@ -311,275 +350,6 @@ export function writeChart(chart: SheetChart, sheetName: string): ChartWriteResu
 
 // ── Title ────────────────────────────────────────────────────────────
 
-function buildTitle(
-  title: string,
-  overlay: boolean,
-  rotationDeg: number | undefined,
-  fontSizePt: number | undefined,
-  bold: boolean | undefined,
-  italic: boolean | undefined,
-  rgbHex: string | undefined,
-  strike: boolean | undefined,
-  underline: boolean | undefined,
-  fontFamily: string | undefined,
-  layout: ResolvedManualLayout | undefined,
-  fillRgbHex: string | undefined,
-  borderRgbHex: string | undefined,
-  borderWidthPt: number | undefined,
-  borderDash: ChartBorderDash | undefined,
-): string {
-  // OOXML's `<a:bodyPr rot="N"/>` attribute is in 60000ths of a degree.
-  // The writer holds `titleRotation` in whole degrees and converts at
-  // emit time. Absence (`undefined`) collapses to the OOXML default
-  // `0` so a fresh chart matches Excel's reference serialization
-  // byte-for-byte.
-  const rot = rotationDeg === undefined ? 0 : rotationDeg * TITLE_ROT_PER_DEGREE;
-  // OOXML's `<a:defRPr sz="N"/>` / `<a:rPr sz="N"/>` attribute is in
-  // 100ths of a point. The writer holds `titleFontSize` in points and
-  // converts at emit time. Absence (`undefined`) collapses to the
-  // application-default `1400` (14pt) so a fresh chart matches Excel's
-  // reference serialization byte-for-byte. The size lands on both the
-  // default-paragraph `<a:defRPr>` and the literal run's `<a:rPr>` so
-  // a re-parse picks the value up off either canonical slot.
-  const sz =
-    fontSizePt === undefined ? TITLE_DEFAULT_FONT_SIZE_SZ : fontSizePt * TITLE_FONT_SZ_PER_POINT;
-  // OOXML's `<a:defRPr b=".."/>` / `<a:rPr b=".."/>` attribute is the
-  // `xsd:boolean` bold flag on `CT_TextCharacterProperties`. The writer
-  // holds `titleBold` as a boolean and emits `1` / `0` at the canonical
-  // slots. Absence (`undefined`) collapses to the OOXML default `0`
-  // (non-bold) so a fresh chart matches Excel's reference serialization
-  // byte-for-byte. Like the size, the flag lands on both the
-  // default-paragraph `<a:defRPr>` and the literal run's `<a:rPr>` so
-  // a re-parse picks the value up off either canonical slot — Excel
-  // keeps the two attributes in sync.
-  const b = bold ? 1 : 0;
-  // OOXML's `<a:defRPr i=".."/>` / `<a:rPr i=".."/>` attribute is the
-  // `xsd:boolean` italic flag on `CT_TextCharacterProperties`. Mirrors
-  // the bold pattern: `titleItalic` lands on both the default-paragraph
-  // `<a:defRPr>` and the literal run's `<a:rPr>` so a re-parse picks
-  // the value up off either canonical slot — Excel keeps the two
-  // attributes in sync. Absence (`undefined`) and explicit `false` both
-  // collapse to omitting the attribute so a fresh chart matches Excel's
-  // reference serialization byte-for-byte (Excel itself omits `i` when
-  // the title is non-italic — only the bold flag is always emitted).
-  const i = italic === true ? 1 : undefined;
-  // OOXML's `<a:defRPr strike=".."/>` / `<a:rPr strike=".."/>` attribute
-  // is the `ST_TextStrikeType` enum on `CT_TextCharacterProperties` —
-  // `"noStrike"` (default), `"sngStrike"` (single line, the value
-  // Excel's UI emits), `"dblStrike"` (double line, non-UI). The writer
-  // emits only the UI variant `"sngStrike"` to keep the surfaced shape
-  // consistent with what Excel's reference UI authors. Absence
-  // (`undefined`) and explicit `false` both collapse to omitting the
-  // attribute (Excel itself omits `strike` when the title is not
-  // strikethrough — the OOXML default `"noStrike"` collapses to
-  // absence). Like bold / italic, the value lands on both the
-  // default-paragraph `<a:defRPr>` and the literal run's `<a:rPr>` so
-  // a re-parse picks the value up off either canonical slot — Excel
-  // keeps the two attributes in sync.
-  const strikeAttr = strike === true ? "sngStrike" : undefined;
-  // OOXML's `<a:defRPr u=".."/>` / `<a:rPr u=".."/>` attribute is the
-  // `ST_TextUnderlineType` enum on `CT_TextCharacterProperties` —
-  // eighteen values total, with `"none"` as the OOXML default,
-  // `"sng"` as the value Excel's UI authors for the "Underline"
-  // checkbox (single line), `"dbl"` for the non-UI double-line
-  // variant, and sixteen exotic types Excel does not surface. The
-  // writer emits only the UI variant `"sng"` to keep the surfaced
-  // shape consistent with what Excel's reference UI authors. Absence
-  // (`undefined`) and explicit `false` both collapse to omitting the
-  // attribute (Excel itself omits `u` when the title is not
-  // underlined — the OOXML default `"none"` collapses to absence).
-  // Like bold / italic / strike, the value lands on both the
-  // default-paragraph `<a:defRPr>` and the literal run's `<a:rPr>` so
-  // a re-parse picks the value up off either canonical slot — Excel
-  // keeps the two attributes in sync.
-  const underlineAttr = underline === true ? "sng" : undefined;
-  // OOXML's `<a:defRPr><a:solidFill><a:srgbClr val="RRGGBB"/>
-  // </a:solidFill></a:defRPr>` carries the title's font color. The
-  // writer holds `titleColor` as a 6-character uppercase hex string
-  // and lands the `<a:solidFill>` block on both the default-paragraph
-  // `<a:defRPr>` and the literal run's `<a:rPr>` so a re-parse picks
-  // the value up off either canonical slot. Absence (`undefined`)
-  // collapses to omitting the entire `<a:solidFill>` block so the
-  // title inherits the theme text color (Excel's reference behavior
-  // for a fresh chart title that has not had a custom color picked).
-  const solidFillChild = rgbHex
-    ? xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: rgbHex })])
-    : undefined;
-  // OOXML's `<a:defRPr><a:latin typeface=".."/></a:defRPr>` carries the
-  // title's font family. The writer holds `titleFontFamily` as a non-
-  // empty string and lands the `<a:latin>` element on both the default-
-  // paragraph `<a:defRPr>` and the literal run's `<a:rPr>` so a re-
-  // parse picks the typeface up off either canonical slot. Absence
-  // (`undefined`) collapses to omitting the entire `<a:latin>` element
-  // so the title inherits the theme typeface (Excel's reference
-  // behavior for a fresh chart title that has not had a custom font
-  // picked). The `<a:latin>` element follows `<a:solidFill>` per the
-  // CT_TextCharacterProperties child sequence (ECMA-376 Part 1,
-  // §21.1.2.3.7) so a fresh chart with both color and family matches
-  // Excel's reference serialization byte-for-byte.
-  const latinChild = fontFamily ? xmlSelfClose("a:latin", { typeface: fontFamily }) : undefined;
-  // When a fill color or a typeface is set the `<a:defRPr>` /
-  // `<a:rPr>` slots expand from self-closing to wrapping the children;
-  // otherwise the writer keeps the existing self-closing form so a
-  // fresh chart with no custom color or font matches Excel's reference
-  // serialization byte-for-byte. Children are emitted in CT_TextChar
-  // acterProperties' canonical schema order: solidFill first, then
-  // latin.
-  const rPrChildren: string[] = [];
-  if (solidFillChild) rPrChildren.push(solidFillChild);
-  if (latinChild) rPrChildren.push(latinChild);
-  const defRPr =
-    rPrChildren.length > 0
-      ? xmlElement("a:defRPr", { sz, b, i, u: underlineAttr, strike: strikeAttr }, rPrChildren)
-      : xmlSelfClose("a:defRPr", { sz, b, i, u: underlineAttr, strike: strikeAttr });
-  const rPr =
-    rPrChildren.length > 0
-      ? xmlElement(
-          "a:rPr",
-          { lang: "en-US", sz, b, i, u: underlineAttr, strike: strikeAttr },
-          rPrChildren,
-        )
-      : xmlSelfClose("a:rPr", {
-          lang: "en-US",
-          sz,
-          b,
-          i,
-          u: underlineAttr,
-          strike: strikeAttr,
-        });
-  // CT_Title (ECMA-376 Part 1, §21.2.2.210) places the optional
-  // `<c:layout>` between `<c:tx>` and `<c:overlay>`. The writer skips
-  // emission entirely when the caller pinned no coordinates so a fresh
-  // chart matches Excel's reference serialization byte-for-byte (Excel
-  // itself omits the block when the title renders at the auto-layout
-  // position above the plot area). Each axis is independently optional
-  // so the helper drops `<c:x>` / `<c:y>` / `<c:w>` / `<c:h>` slots
-  // whose value did not survive normalization.
-  const layoutXml = buildManualLayout(layout);
-  const titleChildren: string[] = [
-    xmlElement("c:tx", undefined, [
-      xmlElement("c:rich", undefined, [
-        xmlElement(
-          "a:bodyPr",
-          {
-            rot,
-            spcFirstLastPara: 1,
-            vertOverflow: "ellipsis",
-            wrap: "square",
-            anchor: "ctr",
-            anchorCtr: 1,
-          },
-          [],
-        ),
-        xmlSelfClose("a:lstStyle"),
-        xmlElement("a:p", undefined, [
-          xmlElement("a:pPr", undefined, [defRPr]),
-          xmlElement("a:r", undefined, [rPr, xmlElement("a:t", undefined, xmlEscape(title))]),
-        ]),
-      ]),
-    ]),
-  ];
-  if (layoutXml !== undefined) {
-    titleChildren.push(layoutXml);
-  }
-  titleChildren.push(xmlSelfClose("c:overlay", { val: overlay ? 1 : 0 }));
-  // CT_Title (ECMA-376 Part 1, §21.2.2.210) places the optional
-  // `<c:spPr>` between `<c:overlay>` and `<c:txPr>` / `<c:extLst>`.
-  // The writer skips emission entirely when the caller did not pin a
-  // fill or border color so a fresh chart matches Excel's reference
-  // serialization byte-for-byte — Excel itself omits the block
-  // whenever the title renders at the theme defaults (typically a
-  // transparent title background with no visible border, no
-  // `<c:spPr>` block). Authors `<a:solidFill>` for the fill and
-  // `<a:ln>` for the stroke in CT_ShapeProperties schema order;
-  // other CT_ShapeProperties children (effects, gradient / pattern /
-  // picture fills, line dash / width / compound styles) are not
-  // modelled at this layer. Distinct from the `<a:defRPr><a:solidFill>`
-  // font-color slot inside `<c:tx><c:rich>` that
-  // {@link SheetChart.titleColor} pins — the typography knobs target
-  // different children of `<c:title>` so a caller can pin both
-  // without conflict.
-  const titleSpPrXml = buildTitleSpPr(fillRgbHex, borderRgbHex, borderWidthPt, borderDash);
-  if (titleSpPrXml !== undefined) {
-    titleChildren.push(titleSpPrXml);
-  }
-  return xmlElement("c:title", undefined, titleChildren);
-}
-
-/**
- * Build the `<c:spPr>` element on `<c:title>` that carries the
- * title's background fill ({@link SheetChart.titleFillColor}),
- * border-stroke color ({@link SheetChart.titleBorderColor}), and
- * border width ({@link SheetChart.titleBorderWidth}). Returns
- * `undefined` when no knob is pinned so the caller can elide the
- * entire block — Excel's reference serialization omits `<c:spPr>`
- * from `<c:title>` whenever the title renders at the theme default
- * fill / stroke (typically a transparent title background with no
- * visible border).
- *
- * When at least one knob lands on the wire, the children are emitted
- * in `CT_ShapeProperties` (ECMA-376 Part 1, §20.1.2.3.13) schema
- * order: `<a:solidFill>` (fill) first, then `<a:ln>` (line / stroke).
- * The fill block has the form `<a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill>`; the stroke block has the form `<a:ln w="EMU">
- * <a:solidFill><a:srgbClr val="RRGGBB"/></a:solidFill></a:ln>`. The
- * `val` attribute holds the canonical 6-character uppercase hex form
- * (the writer normalizes the inputs ahead of this call so malformed
- * source values never reach emit). The width attribute lands on
- * `<a:ln>` (EMU; 1 pt = 12 700 EMU) authored together with the
- * border-color child so a stroke-only or color-only title still emits
- * a single `<a:ln>` block.
- *
- * Mirrors the plot-area / legend `<c:spPr>` slots so a single hex
- * string threads cleanly through every fill / stroke knob the writer
- * authors.
- */
-function buildTitleSpPr(
-  fillRgbHex: string | undefined,
-  borderRgbHex: string | undefined,
-  borderWidthPt: number | undefined,
-  borderDash: ChartBorderDash | undefined,
-): string | undefined {
-  if (
-    fillRgbHex === undefined &&
-    borderRgbHex === undefined &&
-    borderWidthPt === undefined &&
-    borderDash === undefined
-  ) {
-    return undefined;
-  }
-  const children: string[] = [];
-  if (fillRgbHex !== undefined) {
-    children.push(
-      xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fillRgbHex })]),
-    );
-  }
-  if (borderRgbHex !== undefined || borderWidthPt !== undefined || borderDash !== undefined) {
-    const lnAttrs: Record<string, string | number> = {};
-    if (borderWidthPt !== undefined) {
-      // OOXML stores stroke width in EMU (1 pt = 12 700 EMU). Round to
-      // the nearest integer because the schema types `w` as `xsd:int`.
-      lnAttrs.w = Math.round(borderWidthPt * EMU_PER_PT);
-    }
-    const lnChildren: string[] = [];
-    if (borderRgbHex !== undefined) {
-      lnChildren.push(
-        xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: borderRgbHex })]),
-      );
-    }
-    // `<a:prstDash>` follows `<a:solidFill>` per CT_LineProperties
-    // schema sequence (ECMA-376 Part 1, §20.1.2.3.24).
-    if (borderDash !== undefined) {
-      lnChildren.push(xmlSelfClose("a:prstDash", { val: borderDash }));
-    }
-    children.push(
-      lnChildren.length === 0
-        ? xmlSelfClose("a:ln", lnAttrs)
-        : xmlElement("a:ln", Object.keys(lnAttrs).length > 0 ? lnAttrs : undefined, lnChildren),
-    );
-  }
-  return xmlElement("c:spPr", undefined, children);
-}
 
 /**
  * OOXML's `<a:bodyPr rot="N"/>` attribute is in 60000ths of a degree —
@@ -587,48 +357,17 @@ function buildTitleSpPr(
  * emit time. Excel's UI exposes the `-90..90` band; out-of-band values
  * clamp to the nearest endpoint so a corrupt template cannot leak
  * through to the writer either.
- */
-const TITLE_ROT_PER_DEGREE = 60000;
-const TITLE_ROTATION_MIN_DEG = -90;
-const TITLE_ROTATION_MAX_DEG = 90;
-
-/**
- * Normalize a {@link SheetChart.titleRotation} value (whole degrees)
- * for the `<c:title><c:tx><c:rich><a:bodyPr rot="N"/></c:rich></c:tx>
- * </c:title>` writer slot. Returns `undefined` when the input is unset,
- * non-finite, non-numeric, or resolves to `0` after rounding — every
- * absence path collapses to the same omit-the-attribute shape so
- * absence and the OOXML default `0` round-trip identically through
- * {@link cloneChart}. Out-of-range inputs clamp to the `-90..90` band
- * Excel's UI exposes; non-integer inputs round to the nearest whole
- * degree (the OOXML attribute is an integer in 60000ths of a degree,
- * so a fractional whole-degree value has no meaningful refinement at
- * emit time).
- */
-function normalizeTitleRotation(value: number | undefined): number | undefined {
-  if (value === undefined || typeof value !== "number" || !Number.isFinite(value)) return undefined;
-  let degrees = Math.round(value);
-  if (degrees < TITLE_ROTATION_MIN_DEG) degrees = TITLE_ROTATION_MIN_DEG;
-  else if (degrees > TITLE_ROTATION_MAX_DEG) degrees = TITLE_ROTATION_MAX_DEG;
-  if (degrees === 0) return undefined;
-  return degrees;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:bodyPr rot="N"/></c:rich></c:tx>
- * </c:title>` from {@link SheetChart.titleRotation}.
  *
- * Returns the rotation in whole degrees, or `undefined` when the chart
- * leaves the field unset / pinned the OOXML default `0` / passed a
- * non-numeric or non-finite token. The flag is only meaningful when
- * the chart actually emits a title — the caller is expected to gate
- * the call on `showTitle && chart.title`. A chart whose title is
- * suppressed has no `<c:title>` block to host the rotation in either
- * case.
+ * Aliased onto the shared {@link TXPR_ROT_PER_DEGREE} /
+ * {@link ROTATION_MIN_DEG} / {@link ROTATION_MAX_DEG} constants in
+ * `chart/text` so every typography host (chart-title, axis-title,
+ * tick-label, legend, data-label, data-table) shares the same conversion
+ * factor.
  */
-function resolveTitleRotation(chart: SheetChart): number | undefined {
-  return normalizeTitleRotation(chart.titleRotation);
-}
+const TITLE_ROT_PER_DEGREE = TXPR_ROT_PER_DEGREE;
+const TITLE_ROTATION_MIN_DEG = ROTATION_MIN_DEG;
+const TITLE_ROTATION_MAX_DEG = ROTATION_MAX_DEG;
+
 
 /**
  * OOXML's `<a:defRPr sz="N"/>` / `<a:rPr sz="N"/>` attribute is in
@@ -638,10 +377,14 @@ function resolveTitleRotation(chart: SheetChart): number | undefined {
  * writer's clamp uses the same range converted to points (`1..400`pt),
  * so any out-of-range value drops at emit time rather than surface a
  * token Excel would reject.
+ *
+ * Aliased onto the shared {@link FONT_SZ_PER_POINT} /
+ * {@link FONT_SIZE_MIN_PT} / {@link FONT_SIZE_MAX_PT} constants in
+ * `chart/text` so every typography host shares the same range.
  */
-const TITLE_FONT_SZ_PER_POINT = 100;
-const TITLE_FONT_SIZE_MIN_PT = 1;
-const TITLE_FONT_SIZE_MAX_PT = 400;
+const TITLE_FONT_SZ_PER_POINT = FONT_SZ_PER_POINT;
+const TITLE_FONT_SIZE_MIN_PT = FONT_SIZE_MIN_PT;
+const TITLE_FONT_SIZE_MAX_PT = FONT_SIZE_MAX_PT;
 
 /**
  * Application-default `sz` value for the chart title's `<a:defRPr>` /
@@ -652,377 +395,6 @@ const TITLE_FONT_SIZE_MAX_PT = 400;
  */
 const TITLE_DEFAULT_FONT_SIZE_SZ = 1400;
 
-/**
- * Normalize a {@link SheetChart.titleFontSize} value (whole / half
- * points) for the `<c:title><c:tx><c:rich><a:p><a:pPr>
- * <a:defRPr sz="N"/></a:pPr></a:p></c:rich></c:tx></c:title>` writer
- * slot. Returns `undefined` when the input is unset, non-finite,
- * non-numeric, or out of the `1..400`pt band the OOXML
- * `ST_TextFontSize` schema exposes — every absence path collapses to
- * the same default-the-attribute shape so absence and an out-of-range
- * input both fall back to Excel's reference 14pt.
- *
- * Fractional inputs round to the nearest 0.5pt (the OOXML attribute is
- * an integer in 100ths of a point and Excel's UI exposes the same
- * 0.5pt granularity, so finer fractions have no meaningful refinement
- * at emit time).
- */
-function normalizeTitleFontSize(value: number | undefined): number | undefined {
-  if (value === undefined || typeof value !== "number" || !Number.isFinite(value)) return undefined;
-  // Round to the nearest 0.5pt (Excel's UI granularity). `Math.round`
-  // on `2 * value` and dividing by 2 gives a clean half-step band.
-  const halfSteps = Math.round(value * 2);
-  const points = halfSteps / 2;
-  if (points < TITLE_FONT_SIZE_MIN_PT || points > TITLE_FONT_SIZE_MAX_PT) return undefined;
-  return points;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr sz="N"/>
- * </a:pPr></a:p></c:rich></c:tx></c:title>` from
- * {@link SheetChart.titleFontSize}.
- *
- * Returns the size in points (`1..400`), or `undefined` when the chart
- * leaves the field unset / passed an out-of-range or non-numeric
- * token. The flag is only meaningful when the chart actually emits a
- * title — the caller is expected to gate the call on
- * `showTitle && chart.title`. A chart whose title is suppressed has
- * no `<c:title>` block to host the size in either case.
- */
-function resolveTitleFontSize(chart: SheetChart): number | undefined {
-  return normalizeTitleFontSize(chart.titleFontSize);
-}
-
-/**
- * Normalize a {@link SheetChart.titleBold} value for the
- * `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr b=".."/></a:pPr></a:p>
- * </c:rich></c:tx></c:title>` writer slot. Returns the literal
- * boolean when the input is `true` / `false`, or `undefined` for any
- * other token (including `null`-shaped escapes from an untyped
- * caller). Absence and non-boolean tokens both collapse to
- * `undefined` so the writer falls back to the OOXML default `b="0"`
- * (non-bold) Excel itself emits on a fresh chart title.
- */
-function normalizeTitleBold(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr b=".."/>
- * </a:pPr></a:p></c:rich></c:tx></c:title>` from
- * {@link SheetChart.titleBold}.
- *
- * Returns the literal boolean, or `undefined` when the chart leaves
- * the field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a title — the caller is
- * expected to gate the call on `showTitle && chart.title`. A chart
- * whose title is suppressed has no `<c:title>` block to host the flag
- * in either case.
- */
-function resolveTitleBold(chart: SheetChart): boolean | undefined {
-  return normalizeTitleBold(chart.titleBold);
-}
-
-/**
- * Normalize a {@link SheetChart.titleItalic} value for the
- * `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr i=".."/></a:pPr></a:p>
- * </c:rich></c:tx></c:title>` writer slot. Returns the literal
- * boolean when the input is `true` / `false`, or `undefined` for any
- * other token (including `null`-shaped escapes from an untyped
- * caller). Absence and non-boolean tokens both collapse to
- * `undefined` so the writer omits the `i` attribute (Excel's reference
- * serialization for a non-italic title — the OOXML default `false`
- * collapses to absence).
- */
-function normalizeTitleItalic(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr i=".."/>
- * </a:pPr></a:p></c:rich></c:tx></c:title>` from
- * {@link SheetChart.titleItalic}.
- *
- * Returns the literal boolean, or `undefined` when the chart leaves
- * the field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a title — the caller is
- * expected to gate the call on `showTitle && chart.title`. A chart
- * whose title is suppressed has no `<c:title>` block to host the flag
- * in either case.
- */
-function resolveTitleItalic(chart: SheetChart): boolean | undefined {
-  return normalizeTitleItalic(chart.titleItalic);
-}
-
-/**
- * Normalize a {@link SheetChart.titleColor} value for the
- * `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:rich></c:tx></c:title>` writer slot. Returns the 6-character
- * uppercase hex form when the input is a valid sRGB triple (with or
- * without a leading `#`), or `undefined` for any malformed token —
- * wrong length, non-hex characters, alpha-channel forms, or
- * non-string escapes from an untyped caller.
- *
- * Absence and malformed tokens both collapse to `undefined` so the
- * writer skips the entire `<a:solidFill>` block and the title
- * inherits the theme text color (Excel's reference behavior for a
- * fresh chart title without a custom color).
- */
-function normalizeTitleColor(value: string | undefined): string | undefined {
-  return normalizeRgbHexShared(value);
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:rich></c:tx></c:title>` from {@link SheetChart.titleColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed a
- * malformed token. The fill is only meaningful when the chart
- * actually emits a title — the caller is expected to gate the call
- * on `showTitle && chart.title`. A chart whose title is suppressed
- * has no `<c:title>` block to host the fill in either case.
- */
-function resolveTitleColor(chart: SheetChart): string | undefined {
-  return normalizeTitleColor(chart.titleColor);
-}
-
-/**
- * Resolve `<c:title><c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></c:spPr></c:title>` from
- * {@link SheetChart.titleFillColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title font
- * color / plot-area fill / legend fill resolvers exactly. The fill
- * is only meaningful when the chart actually emits a title — the
- * caller is expected to gate the call on `showTitle && chart.title`.
- * A chart whose title is suppressed has no `<c:title>` block to host
- * the `<c:spPr>` slot in either case.
- *
- * Independent of {@link resolveTitleColor}: the fill lands on
- * `<c:title><c:spPr>`, the font color lands on the
- * `<a:defRPr><a:solidFill>` slot inside `<c:tx><c:rich><a:p><a:pPr>`
- * — the two resolvers target different children of `<c:title>` so a
- * single configuration call can pin both.
- */
-function resolveTitleFillColor(chart: SheetChart): string | undefined {
-  return normalizeTitleColor(chart.titleFillColor);
-}
-
-/**
- * Resolve `<c:title><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:title>` from
- * {@link SheetChart.titleBorderColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches every other
- * `<a:srgbClr>` slot the writer authors. The stroke is only
- * meaningful when the chart actually emits a title — the caller is
- * expected to gate the call on `showTitle && chart.title`. A chart
- * whose title is suppressed has no `<c:title>` block to host the
- * `<c:spPr>` slot in either case.
- *
- * Independent of {@link resolveTitleFillColor}: the stroke lands on
- * `<c:title><c:spPr><a:ln>`, the fill lands on
- * `<c:title><c:spPr><a:solidFill>` — the two resolvers target
- * different children of the shared `<c:spPr>` block so a single
- * configuration call can pin both. Mirrors
- * {@link normalizePlotAreaBorderColor} — same hex grammar, distinct
- * host element (`<c:title>` vs `<c:plotArea>`).
- */
-function resolveTitleBorderColor(chart: SheetChart): string | undefined {
-  return normalizeTitleColor(chart.titleBorderColor);
-}
-
-/**
- * Resolve `<c:title><c:spPr><a:ln w="EMU"/></c:spPr></c:title>` from
- * {@link SheetChart.titleBorderWidth}.
- *
- * Returns the point value clamped to the `0.25..13.5` pt band Excel's
- * UI exposes and snapped to the 0.25 pt grid, or `undefined` when the
- * chart leaves the field unset / passed a malformed token (`NaN`,
- * `Infinity`, non-finite). Delegates to {@link clampStrokeWidthPt} so
- * the snap / clamp grammar matches every other `<a:ln w=..>` slot the
- * writer authors (the series stroke knob `series[i].stroke.width`,
- * the plot-area border width knob {@link SheetChart.plotAreaBorderWidth},
- * and the legend border width knob {@link SheetChart.legendBorderWidth}).
- * The width is only meaningful when the chart actually emits a
- * title — the caller is expected to gate the call on
- * `showTitle && chart.title`. A chart whose title is suppressed has no
- * `<c:title>` block to host the `<c:spPr>` slot in either case.
- *
- * Independent of {@link resolveTitleBorderColor}: both knobs land on
- * the same `<a:ln>` element but on a different slot (the color child
- * `<a:solidFill>` versus the line's `w` attribute). Mirrors the
- * plot-area / legend `<c:spPr>` slots — same EMU encoding, same
- * `<a:ln>` host — but lands on `<c:title>`'s own `<c:spPr>` block.
- */
-function resolveTitleBorderWidth(chart: SheetChart): number | undefined {
-  return clampStrokeWidthPt(chart.titleBorderWidth);
-}
-
-/**
- * Resolve `<c:title><c:spPr><a:ln><a:prstDash val=".."/></a:ln></c:spPr>
- * </c:title>` from {@link SheetChart.titleBorderDash}.
- *
- * Returns the recognized {@link ChartBorderDash} value, or `undefined`
- * for the OOXML default `"solid"` and every unrecognized token —
- * delegates to {@link normalizeBorderDash} so the accept / drop grammar
- * matches every other `<a:prstDash>` slot the writer authors. The
- * caller is expected to gate the call on `showTitle && chart.title`
- * since a chart whose title is suppressed has no `<c:title>` block to
- * host the `<c:spPr>` slot.
- *
- * Independent of {@link resolveTitleBorderColor} and
- * {@link resolveTitleBorderWidth}: all three knobs land on the same
- * `<a:ln>` element but on different children / attributes — color is
- * `<a:solidFill>`, width is the `w` attribute, dash is `<a:prstDash>`.
- */
-function resolveTitleBorderDash(chart: SheetChart): ChartBorderDash | undefined {
-  return normalizeBorderDash(chart.titleBorderDash);
-}
-
-/**
- * Normalize a {@link SheetChart.titleStrike} value for the
- * `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr strike=".."/></a:pPr>
- * </a:p></c:rich></c:tx></c:title>` writer slot. Returns the literal
- * boolean when the input is `true` / `false`, or `undefined` for any
- * other token (including `null`-shaped escapes from an untyped
- * caller). Absence and non-boolean tokens both collapse to
- * `undefined` so the writer omits the `strike` attribute entirely
- * (Excel's reference serialization for a non-strikethrough title —
- * the OOXML default `"noStrike"` collapses to absence; only an
- * explicit `true` emits `strike="sngStrike"`).
- */
-function normalizeTitleStrike(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr strike=".."/>
- * </a:pPr></a:p></c:rich></c:tx></c:title>` from
- * {@link SheetChart.titleStrike}.
- *
- * Returns the literal boolean, or `undefined` when the chart leaves
- * the field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a title — the caller is
- * expected to gate the call on `showTitle && chart.title`. A chart
- * whose title is suppressed has no `<c:title>` block to host the flag
- * in either case.
- */
-function resolveTitleStrike(chart: SheetChart): boolean | undefined {
-  return normalizeTitleStrike(chart.titleStrike);
-}
-
-/**
- * Normalize a {@link SheetChart.titleUnderline} value for the
- * `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr u=".."/></a:pPr>
- * </a:p></c:rich></c:tx></c:title>` writer slot. Returns the literal
- * boolean when the input is `true` / `false`, or `undefined` for any
- * other token (including `null`-shaped escapes from an untyped
- * caller). Absence and non-boolean tokens both collapse to
- * `undefined` so the writer omits the `u` attribute entirely (Excel's
- * reference serialization for a non-underlined title — the OOXML
- * default `"none"` collapses to absence; only an explicit `true`
- * emits `u="sng"`).
- */
-function normalizeTitleUnderline(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr u=".."/>
- * </a:pPr></a:p></c:rich></c:tx></c:title>` from
- * {@link SheetChart.titleUnderline}.
- *
- * Returns the literal boolean, or `undefined` when the chart leaves
- * the field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a title — the caller is
- * expected to gate the call on `showTitle && chart.title`. A chart
- * whose title is suppressed has no `<c:title>` block to host the flag
- * in either case.
- */
-function resolveTitleUnderline(chart: SheetChart): boolean | undefined {
-  return normalizeTitleUnderline(chart.titleUnderline);
-}
-
-/**
- * Normalize a {@link SheetChart.titleFontFamily} value for the
- * `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:rich></c:tx>
- * </c:title>` writer slot. Returns the trimmed typeface string when
- * the input is a non-empty string, or `undefined` for any malformed
- * token — empty / whitespace-only strings, or non-string escapes from
- * an untyped caller (`null`, numbers, booleans, etc.).
- *
- * Absence and malformed tokens both collapse to `undefined` so the
- * writer skips the entire `<a:latin>` element and the title inherits
- * the theme typeface (Excel's reference behavior for a fresh chart
- * title without a custom font picked).
- */
-function normalizeTitleFontFamily(value: string | undefined): string | undefined {
-  if (typeof value !== "string") return undefined;
-  const trimmed = value.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Resolve `<c:title><c:tx><c:rich><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:rich></c:tx></c:title>`
- * from {@link SheetChart.titleFontFamily}.
- *
- * Returns the trimmed typeface string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed an empty
- * or non-string token. The element is only meaningful when the chart
- * actually emits a title — the caller is expected to gate the call
- * on `showTitle && chart.title`. A chart whose title is suppressed
- * has no `<c:title>` block to host the typeface in either case.
- */
-function resolveTitleFontFamily(chart: SheetChart): string | undefined {
-  return normalizeTitleFontFamily(chart.titleFontFamily);
-}
-
-/**
- * Resolve `<c:title><c:overlay val=".."/></c:title>` from
- * {@link SheetChart.titleOverlay}.
- *
- * Defaults to `false` (the OOXML default Excel itself emits — the title
- * reserves its own slot above the plot area and the plot area shrinks
- * to make room). Anything other than literal `true` collapses to `false`
- * so a stray non-boolean leaking through the type guard (e.g. `0` / `1` /
- * `"true"` / `null`) never produces `<c:overlay val="1"/>`. This matches
- * how `legendOverlay` / `roundedCorners` / `plotVisOnly` / axis `hidden`
- * treat their inputs: a literal boolean is the only path to a non-default
- * value.
- *
- * The writer always emits `<c:overlay>` inside `<c:title>` because Excel's
- * reference serialization includes the element on every visible title;
- * only the `val` flips when the caller pins `titleOverlay: true`.
- *
- * The flag is only meaningful when the chart actually emits a title — the
- * caller is expected to gate the call on `showTitle && chart.title`. A
- * chart whose title is suppressed has no `<c:title>` block to host the
- * overlay element.
- */
-function resolveTitleOverlay(chart: SheetChart): boolean {
-  return chart.titleOverlay === true;
-}
 
 /**
  * Resolve `<c:autoTitleDeleted val=".."/>` from
@@ -3004,10 +2376,13 @@ function normalizeAxisHidden(value: boolean | undefined): boolean {
  * time. Excel's UI exposes the `-90..90` band; out-of-band values clamp
  * to the nearest endpoint so a corrupt template cannot leak through to
  * the writer either.
+ *
+ * Aliased onto the shared {@link TXPR_ROT_PER_DEGREE} /
+ * {@link ROTATION_MIN_DEG} / {@link ROTATION_MAX_DEG} constants in
+ * `chart/text` so every typography host shares the same conversion factor.
  */
-const TXPR_ROT_PER_DEGREE = 60000;
-const LABEL_ROTATION_MIN_DEG = -90;
-const LABEL_ROTATION_MAX_DEG = 90;
+const LABEL_ROTATION_MIN_DEG = ROTATION_MIN_DEG;
+const LABEL_ROTATION_MAX_DEG = ROTATION_MAX_DEG;
 
 /**
  * Normalize an axis `labelRotation` value (whole degrees) for the
@@ -6143,21 +5518,6 @@ function resolveLegendOverlay(chart: SheetChart): boolean {
 // ── Manual Layout ────────────────────────────────────────────────────
 
 /**
- * Normalized `<c:manualLayout>` coordinate set after the writer runs
- * the caller's input through the `0..1` range filter. Each axis is
- * independently optional — a caller can pin only the position
- * (`x` / `y`) and let the element keep its automatic size, only the
- * size (`w` / `h`) and let it keep its automatic anchor, or any
- * combination.
- */
-interface ResolvedManualLayout {
-  x?: number;
-  y?: number;
-  w?: number;
-  h?: number;
-}
-
-/**
  * Resolve `<c:legend><c:layout><c:manualLayout>...</c:manualLayout>
  * </c:layout></c:legend>` from {@link SheetChart.legendLayout}.
  *
@@ -6278,30 +5638,6 @@ function resolveLegendBorderDash(chart: SheetChart): ChartBorderDash | undefined
   return normalizeBorderDash(chart.legendBorderDash);
 }
 
-/**
- * Resolve `<c:title><c:layout><c:manualLayout>...</c:manualLayout>
- * </c:layout></c:title>` from {@link SheetChart.titleLayout}.
- *
- * Returns the normalized coordinate set, or `undefined` when every
- * axis the caller pinned dropped to `undefined` (so the writer can
- * elide the entire `<c:layout>` block — Excel's reference serialization
- * omits the element when the title renders at the auto-layout position
- * above the plot area). The element is only meaningful when the chart
- * actually emits a title — the caller is expected to gate the call on
- * the resolved title visibility (showTitle && chart.title).
- *
- * Coordinates outside the OOXML `0..1` band, `NaN`, `Infinity`, and
- * non-numeric inputs all collapse to `undefined` on the matching axis
- * so the writer drops the matching `<c:x>` / `<c:y>` / `<c:w>` /
- * `<c:h>` slot rather than emit a token Excel would reject. Mirrors
- * {@link resolveLegendLayout} — same accept-or-drop grammar, same
- * `ChartManualLayout` shape — so a caller can thread a single layout
- * value through both the chart title and the legend without
- * bookkeeping a second type.
- */
-function resolveTitleLayout(chart: SheetChart): ResolvedManualLayout | undefined {
-  return normalizeManualLayout(chart.titleLayout);
-}
 
 /**
  * Resolve `<c:plotArea><c:layout><c:manualLayout>...</c:manualLayout>
@@ -6323,86 +5659,6 @@ function resolveTitleLayout(chart: SheetChart): ResolvedManualLayout | undefined
  */
 function resolvePlotAreaLayout(chart: SheetChart): ResolvedManualLayout | undefined {
   return normalizeManualLayout(chart.plotAreaLayout);
-}
-
-/**
- * Normalize a {@link ChartManualLayout} into the writer's emit-ready
- * shape. Drops every axis whose input is non-numeric / non-finite /
- * out of the `0..1` band; returns `undefined` when every axis dropped
- * so the caller can elide the entire `<c:layout>` block.
- *
- * The accept-and-clamp grammar matches the OOXML `CT_ManualLayout`
- * schema — `<c:x>` / `<c:y>` / `<c:w>` / `<c:h>` carry `xsd:double`
- * values in the `0..1` band per Excel's reference serialization. The
- * normalizer does not silently clamp out-of-range inputs to the
- * endpoints — it drops them outright, mirroring how `titleFontSize` /
- * `axisTitleFontSize` / `legendFontSize` collapse out-of-range numbers
- * rather than emit a token Excel would reject.
- */
-function normalizeManualLayout(
-  raw: ChartManualLayout | undefined,
-): ResolvedManualLayout | undefined {
-  if (!raw || typeof raw !== "object") return undefined;
-  const out: ResolvedManualLayout = {};
-  const x = normalizeLayoutCoordinate(raw.x);
-  if (x !== undefined) out.x = x;
-  const y = normalizeLayoutCoordinate(raw.y);
-  if (y !== undefined) out.y = y;
-  const w = normalizeLayoutCoordinate(raw.w);
-  if (w !== undefined) out.w = w;
-  const h = normalizeLayoutCoordinate(raw.h);
-  if (h !== undefined) out.h = h;
-  if (out.x === undefined && out.y === undefined && out.w === undefined && out.h === undefined) {
-    return undefined;
-  }
-  return out;
-}
-
-/**
- * Normalize a single `<c:x>` / `<c:y>` / `<c:w>` / `<c:h>` coordinate.
- * Accepts a finite number in the `0..1` band; everything else drops to
- * `undefined`.
- */
-function normalizeLayoutCoordinate(raw: unknown): number | undefined {
-  if (typeof raw !== "number") return undefined;
-  if (!Number.isFinite(raw)) return undefined;
-  if (raw < 0 || raw > 1) return undefined;
-  return raw;
-}
-
-/**
- * Build the `<c:layout><c:manualLayout>...</c:manualLayout></c:layout>`
- * block for a resolved layout. Returns `undefined` when the input is
- * `undefined` so the caller can elide the entire block.
- *
- * The writer always emits the `<c:xMode>` / `<c:yMode>` / `<c:wMode>` /
- * `<c:hMode>` children with `val="edge"` whenever the matching `<c:x>` /
- * `<c:y>` / `<c:w>` / `<c:h>` slot is present — `"edge"` is Excel's
- * reference shape when the user drags an element to a custom position
- * (the coordinates are absolute fractions of the chart frame, not
- * deltas from the auto-layout baseline). The `"factor"` form (delta
- * from auto-layout) is read on parse but normalized to `"edge"` on
- * emit so a re-parse after a clone-through stays canonical.
- *
- * The OOXML `CT_ManualLayout` sequence places the mode children before
- * the value children: `<c:layoutTarget>?` / `<c:xMode>?` / `<c:yMode>?`
- * / `<c:wMode>?` / `<c:hMode>?` / `<c:x>?` / `<c:y>?` / `<c:w>?` /
- * `<c:h>?` (ECMA-376 Part 1, §21.2.2.115). The writer emits in that
- * order so a re-parse sees the canonical shape.
- */
-function buildManualLayout(layout: ResolvedManualLayout | undefined): string | undefined {
-  if (!layout) return undefined;
-  const children: string[] = [];
-  if (layout.x !== undefined) children.push(xmlSelfClose("c:xMode", { val: "edge" }));
-  if (layout.y !== undefined) children.push(xmlSelfClose("c:yMode", { val: "edge" }));
-  if (layout.w !== undefined) children.push(xmlSelfClose("c:wMode", { val: "edge" }));
-  if (layout.h !== undefined) children.push(xmlSelfClose("c:hMode", { val: "edge" }));
-  if (layout.x !== undefined) children.push(xmlSelfClose("c:x", { val: layout.x }));
-  if (layout.y !== undefined) children.push(xmlSelfClose("c:y", { val: layout.y }));
-  if (layout.w !== undefined) children.push(xmlSelfClose("c:w", { val: layout.w }));
-  if (layout.h !== undefined) children.push(xmlSelfClose("c:h", { val: layout.h }));
-  if (children.length === 0) return undefined;
-  return xmlElement("c:layout", undefined, [xmlElement("c:manualLayout", undefined, children)]);
 }
 
 // ── Display Blanks As ────────────────────────────────────────────────
