+++ src/xlsx/chart/index.ts
@@ -14,3 +14,4 @@ export * from "./text";
 export * from "./layout";
 export * from "./util";
 export * from "./walls";
+export * from "./title";
